Cleaned up entity registration, added first loc text (#5).
The way entities were being registered was not friendly with other mods.
This method has the advantage of not only allowing us to localize the
entity names sanely, but also avoids potential ID conflicts with other
mods that may add fairies, etc.

This IS a breaking change, any existing fairies will unspawn.